setup basic navigation and pages
